separate sharding from main
